Removed Longview iptables rules
Iptables cannot resolve addresses and as such adding iptables rules with domain names won't work. For example part of starting up the network is starting iptables which would require you do a DNS query to find out what longview.linode.com resolves to but because the network isn't up that resolution fails and because the resolution fails loading the iptables rules fails which ultimately causes the networking daemon to fail. All Longview requires is that the Linode accepts related, established traffic in and port 443 out so declaring longview.linode.com is redundant.

This second commit is to update line numbering to reflect the removal.