Updated SSL cert and key config line entries
The format for the location of the SSL cert and key file entries has changed in Dovecot's config file.  I'm not sure which version the change took place in but I'm running version 2.2.10 on Cent OS 7 and had to use the format specified in my change (or Dovecot would fail to respond to pop3 connections).  This is a great guide, thanks.  This was the only "gotcha" I ran into so I wanted to contribute the update.