Update lemp-server-on-ubuntu-16-04.md
SEO (and few copy edits) edits made. If the PHP configuration page image is used in this tutorial, an additional edit is needed. The code below should be inserted into the meta-header. 

Because I am not a dev and am not sure that my coding is accurate, here is my suggestion of what is needed in terms of using alt text for SERP: <img src="http://www.linode.com/docs/docs / assets / lemp-1604-phpinfo-small.png" alt=“LEMP on Ubuntu 16.04 PHP config screenshot">