Update introduction-to-linux-concepts.md
FWIW, I wanted to eradicate the gerund (installing, booting, connecting vs install, boot, connect) for clarity's sake, but I noticed you link throughout other docs to the subheads here, so I didn't want to introduce a lot of maintenance upkeep for a semi-minor clarity quibble.

I'd vote for killing the distributions section, as I don't think it adds anything substantial. Same for the history section, though as I mentioned, it might be beneficial for total n00bs to get a sense of the culture they're falling into. 

It might be helpful to evaluate this guide in the context of the getting started guide, since there is a fair bit of cross-linking that can get kind of confusing. Some audience typing and evaluation might help tie these together (is the getting started guide also aimed at Linux newbies? Will most of the people reading the getting started guide know Linux or not? What parts of Linux are confusing people?) -- with this in mind, it might make sense to rip out the first half of the explanatory content, and spend most of this topic talking about the main commands, how files and directories are organized, and how to determine/set file access permissions (and the organization of those). Think of it in the sense of "If I'm new to Linode, where should I start?" 

I attempted to separate out conceptual information from tutorial content (keeping the steps segregated from the intros) but it could use some extra workshopping. Also because I think adding the steps may have killed the code formatting that was on the snippets. Hard to tell in this preview.