Linode Managed is not a maintenance service
Linode Manged has not been a maintenance service at least for as long as
I've been with Linode (1 year and 1 month as of right now). I'm honestly
not sure if it's ever been one, but I know it's misleading to customers
when the first sentence of our Managed Services guide indicates
otherwise. Since our support team uses this frequently in helping
customers set up these services, it would be best to redact any
references to maintenance as it will cause confusion.