Change in wording and details for 2FA
Our support team requires that the expiration date be visible on the card
image to consider it acceptable. Our documentation should be congruent
with this requirement.

We may have multiple payment cards on file, but the only we accept is
the one we are currently billing from or the latest card. We should
clarify on this.

"Owner of the card" is less clear than "cardholder name" since one
refers to a name and the other refers to... well... the owner which
could be a name, a picture, or maybe even the actual person. We really
just need the card name.

There is no "the" when it comes to a government-issued photo ID. We
accept multiple forms of identification including drivers license and
passport.