Update and rename logging-in-to-coreos-container-linux.md to log-in-to-coreos-container-linux.md
Copy edited once more and revised for better SEO.
Need to know - does docsmith take title tag and assign it to overall header on doc? Portent recommends that title tag be different from the H1 "title" spanning page top. Can we differentiate the two?
If so, leave the title tag alone (as I revised it above)  and insert this as the H1 "title": How to log in to CoreOS Container Linux