A step 3 was referring to a step 4 that I removed
Initially, in that step I was recommending something else (not activating https redirection). Then I removed it because the workaround was no longer necessary and today my brain mysteriously/randomly reminded me that I may have this step 3 referring to something I've removed. 

Thought about removing the reference from 3 but then I thought making this recommendation might actually be a better idea for users that wouldn't know what to do there. It's unimportant what they choose since we'll later setup the redirection from PrestaShop too but it's better if the users are not left wondering, possibly googling about what to do when they get to that step.