Updated Copy to Nextcloud 14 (#2163)
Nothing needed changing in the guide aside from the new version number, all of the steps and screenshots are up to date. Will comment back to the Disqus user with more detailed information about their particular Nextcloud Apps,  which are either already updated or in the process of being updated to Nextcloud 14.