Run multiple Wordpress sites on one Linode - Updated for accuracy (#2183)
* Updated for accuracy

* Update index.md

* Travis fixes

* Change title graphic text

* Tech Edit 2

* Copy Edit

Renamed the doc back to `Set Up Apache to Run Multiple WordPress Sites on a Single Linode` with the URL pointing back to `websites/cms/configure-apache-to-run-multiple-wordpress-sites-on-one-linode/`. This was done under Angel's supervision so that we don't lose traffic to the guide. Reapplied the header image as a result of the change back.