Unembedding shortguides
When I originally created the Data Portability section, I created a
number of shortguides in order to embed relevant instructions into the
new data portability guides. For example:

1. I extracted the Download a Disk over SSH section of the Copying a
Disk over SSH guide and put it into a new copy-disk-over-ssh-short
shortguide
2. I then embedded this shortguide into the content's original Copying a
Disk over SSH guide, and I also embedded it into the new Download Files
from your Linode guide.

This was also done for other instructions: making a database dump with
mysql, and using filezilla to download files.

This proved to be harder to copy edit, as copy edit changes inside a
shortguide can result in unintended changes to the "flow" of the
embedding guide.

This commit is rolling back most of the shortguides in favor of just
linking to the content. To take the example of copying a disk over SSH,
I am:

1. Pasting the content of the copy-disk-over-ssh-short shortguide back
into the Copying a Disk over SSH guide, where it originally lived.
2. Linking to that guide from the Download Files from your Linode guide.