Removing references to hyper threading
From conversation with CAker: it's important that we don't mask the fact that we use hyper threading to offer more vCPUs. We will tell customer this if asked, and we will probably put this information somewhere at some point (like in a details section for a theoretical marketing page devoted to dedicated CPU). But, this information doesn't really help someone choose which Linode plan they should use, so it's out of scope for the guide.