[New] How to Choose a Linode Plan Size  (#2291)
* initial commit

* vale

* Copy Edit 1

* Copy edit and reformatting

* Fix Travis errors

* Changed some wording

* Added two sections at top, expanded some use cases, other copy edits

* Copy Edit, updated dictionary.

* Removing guide at old how-to-choose-a-linode-plan-size slug

* Changed the word plan to instance. Added vcpu to chart

* Specified a bit more re: hyperthreading

* Copy edit; using the word instance more frequently, replacing most examples of core with vCPU

* Removing references to hyper threading

From conversation with CAker: it's important that we don't mask the fact that we use hyper threading to offer more vCPUs. We will tell customer this if asked, and we will probably put this information somewhere at some point (like in a details section for a theoretical marketing page devoted to dedicated CPU). But, this information doesn't really help someone choose which Linode plan they should use, so it's out of scope for the guide.