Changed to click on advanced tab (#2666)
Currently is says to select settings and then select advanced from there.  However, the advanced settings must have been changed to their own tab.  Which I like as it does not bury the options so deep.  Anywho, the picture will also need to be changed so that you can see the advanced tab.