Fixing support for image frontmatter
If a filename for an image was passed to the image frontmatter,
Hugo would create a meta tag whose link only consisted of the
filename.

Absolute filenames are required for og image meta tags. With this
change, an author can pass the filename of an image in their
guide's bundle, and Hugo will generate the corresponding absolute
link.