Adding conditional logic for which Twitter card summary to use
Some conditional logic already existed. A twitter frontmatter
dictionary was being checked which accepted a card type and
inserted that into the head of a guide. This is now
documented in yaml_headers.md, and it has been added to the
Travis frontmatter test.

Adding an else-if clause which checks if a social media image is
set in frontmatter. If there is, then a summary_large_image
card will be displayed. Otherwise, a small summary card will be
displayed.