Support for default social media image (#2734)
* Support for default social media image

* Use absolute URL for og:image

It seems like relative URLs cause problems for Twitter/Facebook

* Fixing support for image frontmatter

If a filename for an image was passed to the image frontmatter,
Hugo would create a meta tag whose link only consisted of the
filename.

Absolute filenames are required for og image meta tags. With this
change, an author can pass the filename of an image in their
guide's bundle, and Hugo will generate the corresponding absolute
link.

* Added a show_social_image_in_article frontmatter field

If true, and if a social media image is set with the image
frontmatter field, then that image will automatically appear at
the top of its guide.

If false, any social media image set will not be displayed at the
top of the guide.

* Adding conditional logic for which Twitter card summary to use

Some conditional logic already existed. A twitter frontmatter
dictionary was being checked which accepted a card type and
inserted that into the head of a guide. This is now
documented in yaml_headers.md, and it has been added to the
Travis frontmatter test.

Adding an else-if clause which checks if a social media image is
set in frontmatter. If there is, then a summary_large_image
card will be displayed. Otherwise, a small summary card will be
displayed.

* Update default og:image, set default twitter:image

- Updated the default file for the og:image tag with a new asset
from the art department.
- Added support for a separate default twitter:image. This is a
smaller, square-aspect image suitable for small Twitter summary
cards. Default image also provided by art department.

* Add social image to Object Storage Use Cases guide