Fixed aliases to redirect to Linode CLI API v4 version.
Currently: Google's top two results for Linode CLI link to both
`platform/api/using-the-linode-cli` and `platform/api/linode-cli`,
respectively.

When an alias exists, hugo will create a directory to
match the alias entry that contains an html file specifying
the canonical URL for the page and a new redirect target.
Since, there is an existing guide in the alias location
(`platform/api/using-the-linode-cli`) aliasing was not working
as expected. Visiting a `platform/api/using-the-linode-cli` still
directed you to the deprecated guide.

For this reason, I renamed the deprecated guide's url to:
`/platform/api/using-the-linode-cli-api-v3/`. I moved all aliases
in the deprecated guide to the current guide. Now, all previous aliases
point to our current version of the guide.

I also changed the deprecated guide's H1 to
"Linode CLI (API v3 - Deprecated)", so it is immediately
visible to users that this guide is *Deprecated*.