deselect image, pipe ISO to sha256sum, sudo req'd
I had problems starting with an arbitrary image and the directions assume selecting no image (I think your people in IRC told me to click the X - the second time I did this tutorial, I did that and had no problems).

We need to sha256sum the iso. It's a NixOS, cultural thing, it's done at every step in the Nix build system, and if you don't have that, I can't recommend this tutorial.

`sudo` will be required to execute `nixos-rebuild switch` - I verified with the nixos guys on discord.