Minor Modifications and added some commands
Replaced YUM with DNF as CentOS 8 is using DNF as a default package manager. Added "sudo ls -dZ /var/www/example.com" to check whether SELinux Context are added to the directory or not. Modified the command "sudo firewall-cmd --zone=public --permanent --add-service={http,https}" which is used to add the given services to firewall using single command this saves time. Added command "sudo firewall-cmd --list-all" to check the list of services and ports which are added to firewall.