Update index.md (#3615)
mysqltuner.com now points to a gun website, after what looks like a recent change:
```
$ whois mysqltuner.com | grep Updat
   Updated Date: 2020-08-01T10:59:45Z
Updated Date: 2020-08-01T10:59:45Z
```

This was brought to our attention by someone on Twitter after we tweeted out a link to this guide: https://twitter.com/zaherg/status/1302611089226706945
```
the link to mysqltuner website is for a gun recommendation blog!! so its better to change it to the github repo instead
```

My change is to link to the GitHub repo, rather than mysqltuner.com, which occurred twice in this doc. There's still a reference to it in the output of the script, but that doesn't link anywhere.