Set slug frontmatter on every guide
Previously, Hugo was using the title of the guide to render guides
under /docs/guides/:slug/. This was sometimes different from the
name of the guide's bundle (i.e. the last part of the filepath for
a given guide). This commit sets the slug frontmatter equal to the
name of the guide's bundle.