Fixes an issue with the title of guides created through Hugo archetypes
When using the "hugo new" command, and when you create a guide as a Hugo
leaf bundle, the title of the new guide would just be set to "Index". This
is because the .File.TranslationBaseName variable that was used would pull
from the "index.md" filename for the new guide. The new change will instead
use the name of the leaf bundle (i.e. the name of the new directory that
contains the guide) to set the guide's title.