Two new scripts that download and upload settings from Algolia
- scripts/download_algolia_settings is designed to download the
settings from the indices in the production Algolia account to
algolia_settings.json. Only the indices that populated by Hugo
are included, so that we don't share the settings for the
linode.com and community site indices.
- scripts/init_algolia_indices uploads settings for each index
that is declared in config.toml, using the downloaded settings for
the indices that are present in algolia_settings.json (or default
settings otherwise). A user can create a free Algolia account
and then use this script to initalize and prepare their account