Add IPv6AcceptRA option for systemd-network config
For static IPv6 addresses the autoconfiguration via SLAAC should not be done.
Hence distros using systemd-network have to be configured appropriately using
the IPv6AcceptRA option.
Otherwise the IP address for the /64 net is still calculated based on the RA
causing this address to be used as main address for outgoing connections.
This is mostly not intended for e.g. mail servers. (spam blocking)