Theme update
This theme update puts in place a short-term fix for a home screen
bug. Previously, the home screen was showing a blank body on
mobile. We identified some code related to touch-swiping behavior
of the carousels on mobile as the culprit.

That code is now commented out, so the home page appears on mobile.
However, the carousels cannot be swiped left/right until the issue
is investigated further.