Add _index.md for top-level bundles page
We need to have this page; otherwise, Hugo will add an entry to
index.json with an empty objectID, and the Algolia upload script
will reject it.

NOTE: I have added noindex: true to this page for now, so Google
doesn't pick it up, as it has no content yet. We should remove
this frontmatter when we start adding guide bundle taxonomies.