Fixes theme 404 issue (including for ref arch content)
This is a hacky way to introduce this change to the turbo module.
The right way to do it is to vendor the fix bep gave us in this PR:
https://github.com/linode/linode-docs-theme/pull/372

But, we ran into a problem when vendoring the theme after that PR
was merged:
https://github.com/linode/linode-docs-theme/issues/375

So, I'm trying out editing the existing file in the docs vendored
turbo module in the meantime till we get issue #375 figured out.